--- app/src/main/java/com/techcourse/ManualHandlerMapping.java
@@ -1,30 +1,0 @@
-package com.techcourse;
-
-import com.interface21.webmvc.servlet.mvc.asis.Controller;
-import com.interface21.webmvc.servlet.mvc.tobe.HandlerMapping;
-import jakarta.servlet.http.HttpServletRequest;
-import java.util.HashMap;
-import java.util.Map;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-public class ManualHandlerMapping implements HandlerMapping {
-
-    private static final Logger log = LoggerFactory.getLogger(ManualHandlerMapping.class);
-
-    private static final Map<String, Controller> controllers = new HashMap<>();
-
-    public void initialize() {
-        log.info("Initialized Handler Mapping!");
-        controllers.keySet()
-                .forEach(path -> log.info("Path : {}, Controller : {}", path, controllers.get(path).getClass()));
-    }
-
-    @Override
-    public Object getHandler(final HttpServletRequest request) {
-        String requestURI = request.getRequestURI();
-        String path = requestURI.substring(request.getContextPath().length());
-        log.debug("Request Mapping Uri : {}", path);
-        return controllers.get(path);
-    }
-}
